Aggrigate issue view
