+++ OpenRedmine/src/main/java/jp/redmine/redmineclient/IssueActivity.java
@@ -20,8 +20,8 @@ import jp.redmine.redmineclient.db.cache.RedmineProjectModel;
 import jp.redmine.redmineclient.entity.RedmineIssue;
 import jp.redmine.redmineclient.entity.RedmineProject;
 import jp.redmine.redmineclient.fragment.ActivityInterface;
-import jp.redmine.redmineclient.fragment.Issue;
 import jp.redmine.redmineclient.fragment.IssueEdit;
+import jp.redmine.redmineclient.fragment.IssueView;
 import jp.redmine.redmineclient.fragment.TimeEntryEdit;
 import jp.redmine.redmineclient.param.IssueArgument;
 import jp.redmine.redmineclient.param.TimeEntryArgument;
@@ -76,7 +76,7 @@ public class IssueActivity extends TabActivity<DatabaseCacheHelper>
 			list.add((new CorePage<IssueArgument>() {
 				@Override
 				public Fragment getRawFragment() {
-					return Issue.newInstance(getParam());
+					return IssueView.newInstance(getParam());
 				}
 
 				@Override
