--- OpenRedmine/src/main/java/jp/redmine/redmineclient/fragment/IssueTitle.java
@@ -1,63 +1,0 @@
-package jp.redmine.redmineclient.fragment;
-
-import java.sql.SQLException;
-
-import com.j256.ormlite.android.apptools.OrmLiteFragment;
-
-import jp.redmine.redmineclient.R;
-import jp.redmine.redmineclient.db.cache.DatabaseCacheHelper;
-import jp.redmine.redmineclient.db.cache.RedmineIssueModel;
-import jp.redmine.redmineclient.entity.RedmineIssue;
-import jp.redmine.redmineclient.form.RedmineIssueViewForm;
-import jp.redmine.redmineclient.param.IssueArgument;
-import android.os.Bundle;
-import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-public class IssueTitle extends OrmLiteFragment<DatabaseCacheHelper> {
-
-	private RedmineIssueViewForm form;
-
-	public IssueTitle(){
-		super();
-	}
-
-	static public IssueTitle newInstance(IssueArgument intent){
-		IssueTitle instance = new IssueTitle();
-		instance.setArguments(intent.getArgument());
-		return instance;
-	}
-
-	@Override
-	public View onCreateView(LayoutInflater inflater, ViewGroup container,
-			Bundle savedInstanceState) {
-		return inflater.inflate(R.layout.issuetitle, container, false);
-	}
-
-	@Override
-	public void onActivityCreated(Bundle savedInstanceState) {
-		super.onActivityCreated(savedInstanceState);
-		form = new RedmineIssueViewForm(getView());
-	}
-
-	@Override
-	public void onStart() {
-		super.onStart();
-
-		IssueArgument intent = new IssueArgument();
-		intent.setArgument(getArguments());
-		int connectionid = intent.getConnectionId();
-
-		RedmineIssueModel model = new RedmineIssueModel(getHelper());
-		RedmineIssue issue = new RedmineIssue();
-		try {
-			issue = model.fetchById(connectionid, intent.getIssueId());
-		} catch (SQLException e) {
-			Log.e("SelectDataTask","ParserIssue",e);
-		}
-
-		form.setValue(issue);
-	}
-}
