--- OpenRedmine/src/main/res/layout/issuecomment.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:validate="http://schemas.android.com/apk/res-auto"
-    android:layout_width="fill_parent"
-    android:layout_height="wrap_content"
-	android:gravity="top"
-    android:orientation="horizontal" >
-	<com.andreabaccega.widget.FormEditText
-	    android:id="@+id/textDescription"
-	    android:hint="@string/comment"
-	    android:layout_width="wrap_content"
-	    android:layout_height="wrap_content"
-	    android:layout_weight="1"
-	    android:ems="10"
-	    android:imeOptions="actionGo"
-	    android:minLines="1"
-	    validate:emptyAllowed="false" />
-
-	<Button
-	    android:id="@+id/buttonOK"
-	    android:layout_width="50dp"
-	    android:layout_height="wrap_content"
-	    android:layout_weight="0"
-	    android:gravity="center_vertical|center_horizontal"
-		android:text="@android:string/ok" />
-</LinearLayout>
