+++ OpenRedmine/src/main/res/layout/issuedetail.xml
@@ -1,0 +1,79 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+	xmlns:validate="http://schemas.android.com/apk/res-auto"
+    android:layout_width="fill_parent"
+    android:layout_height="fill_parent"
+    android:orientation="vertical" >
+	<LinearLayout
+		android:layout_width="fill_parent"
+		android:layout_height="wrap_content"
+		android:orientation="vertical" >
+		<HorizontalScrollView
+			android:id="@+id/layoutTitle1"
+			android:layout_width="wrap_content"
+			android:layout_height="wrap_content" >
+			<LinearLayout
+				android:id="@+id/layoutTitleContent"
+				android:layout_width="wrap_content"
+				android:layout_height="wrap_content"
+				android:orientation="horizontal" >
+				 <TextView
+					 android:id="@+id/textIssueId"
+					 android:layout_width="wrap_content"
+					 android:layout_height="wrap_content"
+					 android:layout_marginLeft="3dip"
+					 android:text="#0000"
+					 android:textAppearance="?android:attr/textAppearanceLarge"
+					 tools:ignore="HardcodedText" />
+
+				 <TextView
+					 android:id="@+id/textSubject"
+					 android:layout_width="wrap_content"
+					 android:layout_height="wrap_content"
+					 android:layout_marginLeft="5dp"
+					 android:text="Issue title here....allow multi-line..."
+					 android:textAppearance="?android:attr/textAppearanceLarge"
+					 tools:ignore="HardcodedText" />
+			</LinearLayout>
+		</HorizontalScrollView>
+		<LinearLayout
+			android:id="@+id/layoutTitle2"
+			android:layout_width="wrap_content"
+			android:layout_height="wrap_content"
+			android:orientation="horizontal"
+			android:visibility="gone" />
+	</LinearLayout>
+	<se.emilsjolander.stickylistheaders.PtrStickyListHeadersListView
+		xmlns:android="http://schemas.android.com/apk/res/android"
+		android:id="@+id/list"
+		android:layout_weight="1"
+		android:layout_width="fill_parent"
+		android:layout_height="0dp"
+		android:orientation="vertical" />
+
+	<LinearLayout
+		android:layout_width="fill_parent"
+		android:layout_height="wrap_content"
+		android:gravity="top"
+		android:orientation="horizontal" >
+		<com.andreabaccega.widget.FormEditText
+			android:id="@+id/textDescription"
+			android:hint="@string/comment"
+			android:layout_width="wrap_content"
+			android:layout_height="wrap_content"
+			android:layout_weight="1"
+			android:ems="10"
+			android:imeOptions="actionGo"
+			android:minLines="1"
+			validate:emptyAllowed="false" />
+
+		<Button
+			android:id="@+id/buttonOK"
+			android:layout_width="50dp"
+			android:layout_height="wrap_content"
+			android:layout_weight="0"
+			android:gravity="center_vertical|center_horizontal"
+			android:text="@android:string/ok" />
+	</LinearLayout>
+</LinearLayout>
