--- OpenRedmine/src/main/res/layout/issuetitle.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="fill_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical" >
-	<HorizontalScrollView
-	    android:id="@+id/layoutTitle1"
-	    android:layout_width="wrap_content"
-	    android:layout_height="wrap_content" >
-		<LinearLayout
-	    	android:id="@+id/layoutTitleContent"
-		    android:layout_width="wrap_content"
-		    android:layout_height="wrap_content"
-		    android:orientation="horizontal" >
-		     <TextView
-		         android:id="@+id/textIssueId"
-		         android:layout_width="wrap_content"
-		         android:layout_height="wrap_content"
-		         android:layout_marginLeft="3dip"
-		         android:text="#0000"
-		         android:textAppearance="?android:attr/textAppearanceLarge"
-		         tools:ignore="HardcodedText" />
-
-		     <TextView
-		         android:id="@+id/textSubject"
-		         android:layout_width="wrap_content"
-		         android:layout_height="wrap_content"
-		         android:layout_marginLeft="5dp"
-		         android:text="Issue title here....allow multi-line..."
-		         android:textAppearance="?android:attr/textAppearanceLarge"
-		         tools:ignore="HardcodedText" />
-		</LinearLayout>
-	</HorizontalScrollView>
-	<LinearLayout
-    	android:id="@+id/layoutTitle2"
-	    android:layout_width="wrap_content"
-	    android:layout_height="wrap_content"
-	    android:orientation="horizontal"
-	    android:visibility="gone" />
-</LinearLayout>
